edit appointment model and add appointment routes
